Workout update 2...

+++ app/src/main/java/com/example/fitnessapplication/WorkoutInterface.java
@@ -2,7 +2,7 @@ package com.example.fitnessapplication;
 
 public interface WorkoutInterface {
 
-    int countCalories();
+    float countCalories();
 
     int calculateSpeed();
